% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramiga-package.R
\docType{package}
\name{ramiga-package}
\alias{ramiga}
\alias{ramiga-package}
\title{ramiga: Emulating the Amiga Using the vAmiga Engine}
\description{
This is an Amiga emulator based on the vAmiga project \url{https://github.com/dirkwhoffmann/vAmiga}. It provides low level access to the emulator and allows users to interact with the emulated machine through the console or scripts.
}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Dirk Hoffmann [copyright holder]
}

}
\keyword{internal}
