% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{RamigaComponent}
\alias{RamigaComponent}
\title{The RamigaComponent R6 Class}
\description{
An abstract class to represent emulator components. Each of the emulator
components inherit from this class. It provides consistent access to and
control over component specific options.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{options}}{A named \code{list} of options. You can get and set
all individual elements. Elements are always returned as
\code{character}, but can be also be set as \code{numeric} or \code{logical},
depending on the specific option. Assignment values are always
parsed by the vAmiga engine.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaComponent-new}{\code{RamigaComponent$new()}}
\item \href{#method-RamigaComponent-get_pointer}{\code{RamigaComponent$get_pointer()}}
\item \href{#method-RamigaComponent-get_emulator}{\code{RamigaComponent$get_emulator()}}
\item \href{#method-RamigaComponent-list_options}{\code{RamigaComponent$list_options()}}
\item \href{#method-RamigaComponent-clone}{\code{RamigaComponent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaComponent-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaComponent-new}{}}}
\subsection{Method \code{new()}}{
Most emulator components don't need to be initialised. It's simplest
to create an emulator with \code{RamigaEmulator$new()}. All components
can be accessed through the R6 object fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaComponent$new(emulator, component_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{All emulator components should be part of an emulator.
All components (except for the emulator itself) therefore needs the,
initialised with an emulator.}

\item{\code{component_id}}{Integer value used to identify different
types of components.}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaComponent-get_pointer"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaComponent-get_pointer}{}}}
\subsection{Method \code{get_pointer()}}{
Get the pointer to the emulator object in memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaComponent$get_pointer(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the \code{externalptr} pointing to the C++ class
instance in memory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaComponent-get_emulator"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaComponent-get_emulator}{}}}
\subsection{Method \code{get_emulator()}}{
Get associated emulator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaComponent$get_emulator(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a \link{RamigaEmulator} class object
instance in memory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaComponent-list_options"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaComponent-list_options}{}}}
\subsection{Method \code{list_options()}}{
List available options for the emulator component. It shows the name
of the option and indicates their intended value(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaComponent$list_options(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a named \code{list} with available options.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaComponent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaComponent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaComponent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
