% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controldevice.R
\name{RamigaControlDevice}
\alias{RamigaControlDevice}
\title{The RamigaControlDevice R6 Class}
\description{
A class that represents control devices, like the mouse or joystick.
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaControlDevice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaControlDevice-new}{\code{RamigaControlDevice$new()}}
\item \href{#method-RamigaControlDevice-print}{\code{RamigaControlDevice$print()}}
\item \href{#method-RamigaControlDevice-clone}{\code{RamigaControlDevice$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlDevice-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlDevice-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlDevice$new(emulator, port_number, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{A virtual control device only is meaningful when connected
to a virtual device. Therefore, you need an emulator (\link{RamigaEmulator})
to connect the device to.}

\item{\code{port_number}}{Port on the virtual machine to which the device is
connected. Should be either port 1 or 2.}

\item{\code{type}}{Type of device, should be either \code{"mouse"}, or \code{"joystick"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlDevice-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlDevice-print}{}}}
\subsection{Method \code{print()}}{
Prints some basic information about the control device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlDevice$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlDevice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlDevice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlDevice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
