% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlport.R
\name{RamigaControlPort}
\alias{RamigaControlPort}
\title{The RamigaControlPort R6 Class}
\description{
A class that represents the emulator's control ports
(for mouse and joystick).
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaControlPort}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{device}}{This field tells you which device (mouse, joystick or none)
is connected to the current port. Should be a \link{RamigaControlDevice}
class object, or \code{NULL} if there is no device.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaControlPort-new}{\code{RamigaControlPort$new()}}
\item \href{#method-RamigaControlPort-print}{\code{RamigaControlPort$print()}}
\item \href{#method-RamigaControlPort-clone}{\code{RamigaControlPort$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlPort-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlPort-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlPort$new(emulator, port_number = 1L, device = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{An \code{RamigaEmulator} class object. The control port to be created needs to be associated with a virtual machine. The initiated object will represent the control port of the emulated machine.}

\item{\code{port_number}}{The Amiga has two control ports. This
number specifies which one we are using. Should be either
1 or 2.}

\item{\code{device}}{Device connected to this port. An \link{RamigaControlDevice}
class object or \code{NULL} if there is no device.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlPort-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlPort-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the control port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlPort$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaControlPort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaControlPort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaControlPort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
