% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.R
\name{RamigaCPU}
\alias{RamigaCPU}
\title{The RamigaCPU R6 Class}
\description{
A class that represents the emulator's Central Processing Unit (CPU).
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaCPU}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaCPU-get_cycles}{\code{RamigaCPU$get_cycles()}}
\item \href{#method-RamigaCPU-get_program_counter}{\code{RamigaCPU$get_program_counter()}}
\item \href{#method-RamigaCPU-get_config}{\code{RamigaCPU$get_config()}}
\item \href{#method-RamigaCPU-print}{\code{RamigaCPU$print()}}
\item \href{#method-RamigaCPU-clone}{\code{RamigaCPU$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="initialize"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-initialize'><code>ramiga::RamigaComponent$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaCPU-get_cycles"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaCPU-get_cycles}{}}}
\subsection{Method \code{get_cycles()}}{
Get the CPU cycle count since the emulator started
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaCPU$get_cycles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the count in number of cycles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaCPU-get_program_counter"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaCPU-get_program_counter}{}}}
\subsection{Method \code{get_program_counter()}}{
Get the memory address of the currently running program
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaCPU$get_program_counter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
On the Amiga memory addresses are mapped as 32 bit
unsigned integers. As R doesn't support this type, it is
returned as a \code{numeric} (double) value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaCPU-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaCPU-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get current CPU configuration details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaCPU$get_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaCPU-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaCPU-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the CPU
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaCPU$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaCPU-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaCPU-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaCPU$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
