% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{RamigaFileSystem}
\alias{RamigaFileSystem}
\title{The RamigaFileSystem R6 Class}
\description{
Obtain the file system from a virtual device (if available), such that
you can interact with it.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaFileSystem-new}{\code{RamigaFileSystem$new()}}
\item \href{#method-RamigaFileSystem-get_volume_name}{\code{RamigaFileSystem$get_volume_name()}}
\item \href{#method-RamigaFileSystem-get_traits}{\code{RamigaFileSystem$get_traits()}}
\item \href{#method-RamigaFileSystem-print}{\code{RamigaFileSystem$print()}}
\item \href{#method-RamigaFileSystem-clone}{\code{RamigaFileSystem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFileSystem-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFileSystem-new}{}}}
\subsection{Method \code{new()}}{
Attempt to access the file system on a device image if available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFileSystem$new(image)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{The virtual device (represented by a \link{RamigaImage}
class object), from which the file system should be retrieved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFileSystem-get_volume_name"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFileSystem-get_volume_name}{}}}
\subsection{Method \code{get_volume_name()}}{
Get the name of the volume of that contains the file system
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFileSystem$get_volume_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the name of the volume
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFileSystem-get_traits"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFileSystem-get_traits}{}}}
\subsection{Method \code{get_traits()}}{
Get some traits of the file system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFileSystem$get_traits(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a named \code{list} of file system traits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFileSystem-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFileSystem-print}{}}}
\subsection{Method \code{print()}}{
Prints basic info about the file system
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFileSystem$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFileSystem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFileSystem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFileSystem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
