% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{RamigaImage}
\alias{RamigaImage}
\title{The RamigaImage R6 Class}
\description{
A virtual device represented by an image.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaImage-new}{\code{RamigaImage$new()}}
\item \href{#method-RamigaImage-info}{\code{RamigaImage$info()}}
\item \href{#method-RamigaImage-size}{\code{RamigaImage$size()}}
\item \href{#method-RamigaImage-get_pointer}{\code{RamigaImage$get_pointer()}}
\item \href{#method-RamigaImage-is_formatted}{\code{RamigaImage$is_formatted()}}
\item \href{#method-RamigaImage-get_file_system}{\code{RamigaImage$get_file_system()}}
\item \href{#method-RamigaImage-print}{\code{RamigaImage$print()}}
\item \href{#method-RamigaImage-clone}{\code{RamigaImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-new}{}}}
\subsection{Method \code{new()}}{
Initialise an image from a disk image file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{File pointing to disk image file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-info"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-info}{}}}
\subsection{Method \code{info()}}{
Get information about the image.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named \code{list} with information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-size"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-size}{}}}
\subsection{Method \code{size()}}{
Get the size of the device represented by the image.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the size of the image in bytes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-get_pointer"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-get_pointer}{}}}
\subsection{Method \code{get_pointer()}}{
Get the pointer to the image object in memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$get_pointer(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the \code{externalptr} pointing to the C++ class
instance in memory.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-is_formatted"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-is_formatted}{}}}
\subsection{Method \code{is_formatted()}}{
Determines if there is a file system present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$is_formatted()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns \code{logical} indicating if disk contains a file system.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-get_file_system"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-get_file_system}{}}}
\subsection{Method \code{get_file_system()}}{
Get a file system from a virtual device if present
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$get_file_system(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a \link{RamigaFileSystem} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
