% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{RamigaMemory}
\alias{RamigaMemory}
\title{The RamigaMemory R6 Class}
\description{
The RamigaMemory class represents the memory of the emulated machine.
Use this object to interact with the machines memory.
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaEmulator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaEmulator-load_rom}{\code{RamigaMemory$load_rom()}}
\item \href{#method-RamigaEmulator-has_rom}{\code{RamigaMemory$has_rom()}}
\item \href{#method-RamigaEmulator-get_rom_traits}{\code{RamigaMemory$get_rom_traits()}}
\item \href{#method-RamigaEmulator-get_config}{\code{RamigaMemory$get_config()}}
\item \href{#method-RamigaEmulator-print}{\code{RamigaMemory$print()}}
\item \href{#method-RamigaEmulator-clone}{\code{RamigaMemory$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="initialize"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-initialize'><code>ramiga::RamigaComponent$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-load_rom"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-load_rom}{}}}
\subsection{Method \code{load_rom()}}{
Load an operating system to ROM. It can be an official Kickstart
ROM image, or the rom provided for AROS Research Operating System.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$load_rom(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{File path pointing to the location of the ROM file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the updated RamigaMemory object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-has_rom"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-has_rom}{}}}
\subsection{Method \code{has_rom()}}{
Check if a ROM image was successfully loaded
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$has_rom()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a \code{logical} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_rom_traits"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_rom_traits}{}}}
\subsection{Method \code{get_rom_traits()}}{
Get ROM characteristics if known.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$get_rom_traits()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns \code{NULL} if no ROM has been loaded. It will return
A named list with ROM traits otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get details on the configuration of the emulated memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$get_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named \code{list} with memory configurations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-print}{}}}
\subsection{Method \code{print()}}{
Print some basic information about the emulated memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
