% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav.R
\name{save_wav}
\alias{save_wav}
\title{Save audio to WAV file}
\usage{
save_wav(samples, path, rate = 44100L)
}
\arguments{
\item{samples}{A \code{matrix} of \code{double} values.
Values need to be between -1 and +1. The matrix needs to have
2 rows. One for each channel.}

\item{path}{Path of the file where the audio needs to be stored}

\item{rate}{Sample rate. Defaults to 44100L}
}
\value{
Nothing
}
\description{
Helper function to write stereo audio to a WAV file.
}
