% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{RamigaEmulator}
\alias{RamigaEmulator}
\title{TODO}
\value{
TODO

TODO

TODO

TODO

TODO
}
\description{
TODO

TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaEmulator-new}{\code{RamigaEmulator$new()}}
\item \href{#method-RamigaEmulator-get_memory}{\code{RamigaEmulator$get_memory()}}
\item \href{#method-RamigaEmulator-get_floppy_drive}{\code{RamigaEmulator$get_floppy_drive()}}
\item \href{#method-RamigaEmulator-start}{\code{RamigaEmulator$start()}}
\item \href{#method-RamigaEmulator-run}{\code{RamigaEmulator$run()}}
\item \href{#method-RamigaEmulator-info}{\code{RamigaEmulator$info()}}
\item \href{#method-RamigaEmulator-print}{\code{RamigaEmulator$print()}}
\item \href{#method-RamigaEmulator-clone}{\code{RamigaEmulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_memory"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_memory}{}}}
\subsection{Method \code{get_memory()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$get_memory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_floppy_drive"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_floppy_drive}{}}}
\subsection{Method \code{get_floppy_drive()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$get_floppy_drive(drive_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-start"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-start}{}}}
\subsection{Method \code{start()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-run"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-run}{}}}
\subsection{Method \code{run()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-info"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-info}{}}}
\subsection{Method \code{info()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-print}{}}}
\subsection{Method \code{print()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
