% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{RamigaEmulator}
\alias{RamigaEmulator}
\title{The RamigaEmulator R6 Class}
\description{
A class representing the emulator for emulating an Amiga.
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaEmulator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{memory}}{The emulated memory, represented by \link{RamigaMemory}}

\item{\code{cpu}}{The emulated CPU, represented by \link{RamigaCPU}}

\item{\code{floppy_drives}}{The emulated floppy drives, represented by
a named \code{list} of \link{RamigaFloppyDrive} objects.}

\item{\code{control_ports}}{Control ports to which a virtual mouse or
joystick can be connected. It is represented by the
\link{RamigaControlPort} class.}

\item{\code{output}}{Get object to capture emulator output. Handled
by \link{RamigaOutput}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaEmulator-new}{\code{RamigaEmulator$new()}}
\item \href{#method-RamigaEmulator-power_on}{\code{RamigaEmulator$power_on()}}
\item \href{#method-RamigaEmulator-power_off}{\code{RamigaEmulator$power_off()}}
\item \href{#method-RamigaEmulator-soft_reset}{\code{RamigaEmulator$soft_reset()}}
\item \href{#method-RamigaEmulator-hard_reset}{\code{RamigaEmulator$hard_reset()}}
\item \href{#method-RamigaEmulator-run}{\code{RamigaEmulator$run()}}
\item \href{#method-RamigaEmulator-next_frame}{\code{RamigaEmulator$next_frame()}}
\item \href{#method-RamigaEmulator-fast_forward}{\code{RamigaEmulator$fast_forward()}}
\item \href{#method-RamigaEmulator-get_info}{\code{RamigaEmulator$get_info()}}
\item \href{#method-RamigaEmulator-get_config}{\code{RamigaEmulator$get_config()}}
\item \href{#method-RamigaEmulator-set_config}{\code{RamigaEmulator$set_config()}}
\item \href{#method-RamigaEmulator-get_version}{\code{RamigaEmulator$get_version()}}
\item \href{#method-RamigaEmulator-print}{\code{RamigaEmulator$print()}}
\item \href{#method-RamigaEmulator-clone}{\code{RamigaEmulator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new Amiga emulator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-power_on"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-power_on}{}}}
\subsection{Method \code{power_on()}}{
When the emulator is initialised it is powered off.
Call this to power on the virtual machine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$power_on()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-power_off"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-power_off}{}}}
\subsection{Method \code{power_off()}}{
Power off emulator. Advisable if you plan to make adjustments
to the emulated hardware configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$power_off()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-soft_reset"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-soft_reset}{}}}
\subsection{Method \code{soft_reset()}}{
A soft reset, just resets the CPU, causing the system to reboot.
This is similar to simultaneously pressing \out{<Ctrl>} and both
\out{<Amiga>} keys on an original machine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$soft_reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-hard_reset"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-hard_reset}{}}}
\subsection{Method \code{hard_reset()}}{
A hard reset. Same as powering of the machine, then powering
it back on.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$hard_reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-run"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-run}{}}}
\subsection{Method \code{run()}}{
When initialised, the emulator is paused. Call this to start
running the virtual machine. It will hold up the R thread
untill you interrupt (by pressing \if{html}{\out{<ESC>}}). After the interrupt
the machine will be paused.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns NULL invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-next_frame"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-next_frame}{}}}
\subsection{Method \code{next_frame()}}{
Update the machines state to the next video frame. The machine
runs until the vertical blank is reached and pauses immediately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$next_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-fast_forward"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-fast_forward}{}}}
\subsection{Method \code{fast_forward()}}{
Runs the emulator at warp speed and skips the specified number
of frames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$fast_forward(frames)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the emulator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get information about the emulator state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$get_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get information about the emulator configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$get_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-set_config"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-set_config}{}}}
\subsection{Method \code{set_config()}}{
Quickly set the emulator configuration to one of the pre-specified
schemes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$set_config(scheme = "A500_OCS_1MB", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scheme}}{Should be one of the following strings:
\code{"A1000_OCS_1MB"}, \code{"A500_OCS_1MB"} (default), \code{"A500_ECS_1MB"},
or \code{"A500_PLUS_1MB"}.}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-get_version}{}}}
\subsection{Method \code{get_version()}}{
Get the version number of the vAmiga core used by this package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$get_version(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-print}{}}}
\subsection{Method \code{print()}}{
Prints basic information about the emulator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaEmulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
