% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floppydrive.R
\name{RamigaFloppyDrive}
\alias{RamigaFloppyDrive}
\title{The RamigaFlopyDrive R6 Class}
\description{
A class that represents the emulator's floppy drive.
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaFloppyDrive}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaFloppyDrive-new}{\code{RamigaFloppyDrive$new()}}
\item \href{#method-RamigaFloppyDrive-insert_disk}{\code{RamigaFloppyDrive$insert_disk()}}
\item \href{#method-RamigaFloppyDrive-eject_disk}{\code{RamigaFloppyDrive$eject_disk()}}
\item \href{#method-RamigaFloppyDrive-get_info}{\code{RamigaFloppyDrive$get_info()}}
\item \href{#method-RamigaFloppyDrive-print}{\code{RamigaFloppyDrive$print()}}
\item \href{#method-RamigaFloppyDrive-clone}{\code{RamigaFloppyDrive$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$new(emulator, drive_number = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{An \code{RamigaEmulator} class object. The floppy drive to be created needs to be associated with a virtual machine. The initiated object will represent the floppy drive of the emulated machine.}

\item{\code{drive_number}}{The emulator has 4 floppy drives, numbered 0 to 3. Pick which drive you want to operate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-insert_disk"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-insert_disk}{}}}
\subsection{Method \code{insert_disk()}}{
Insert a virtual floppy disk in the virtual floppy drive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$insert_disk(disk, write_protected = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disk}}{A floppy disk represented by a \link{RamigaImage} class
object.}

\item{\code{write_protected}}{\code{logical} value indicating whether the
inserted disk needs to be write protected. Default is \code{TRUE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-eject_disk"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-eject_disk}{}}}
\subsection{Method \code{eject_disk()}}{
Eject a virtual floppy disk from the virtual drive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$eject_disk(delay = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delay}}{Delay for ejecting the disk counted in CPU cycles.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get information about the floppy drive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$get_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a named list with information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-print}{}}}
\subsection{Method \code{print()}}{
Prints some basic information about the floppy drive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
