% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floppydrive.R
\name{RamigaFloppyDrive}
\alias{RamigaFloppyDrive}
\title{TODO}
\description{
TODO

TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaFloppyDrive-new}{\code{RamigaFloppyDrive$new()}}
\item \href{#method-RamigaFloppyDrive-insert_disk}{\code{RamigaFloppyDrive$insert_disk()}}
\item \href{#method-RamigaFloppyDrive-eject_disk}{\code{RamigaFloppyDrive$eject_disk()}}
\item \href{#method-RamigaFloppyDrive-print}{\code{RamigaFloppyDrive$print()}}
\item \href{#method-RamigaFloppyDrive-clone}{\code{RamigaFloppyDrive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$new(emulator, drive_number = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{TODO}

\item{\code{drive_number}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-insert_disk"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-insert_disk}{}}}
\subsection{Method \code{insert_disk()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$insert_disk(disk, write_protected = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disk}}{TODO}

\item{\code{write_protected}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-eject_disk"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-eject_disk}{}}}
\subsection{Method \code{eject_disk()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$eject_disk(delay = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delay}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disk}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaFloppyDrive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaFloppyDrive-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaFloppyDrive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
