% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{RamigaImage}
\alias{RamigaImage}
\title{TODO}
\value{
TODO

Returns the size of the image in bytes.

TODO.
}
\description{
TODO

TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaImage-new}{\code{RamigaImage$new()}}
\item \href{#method-RamigaImage-info}{\code{RamigaImage$info()}}
\item \href{#method-RamigaImage-size}{\code{RamigaImage$size()}}
\item \href{#method-RamigaImage-get_pointer}{\code{RamigaImage$get_pointer()}}
\item \href{#method-RamigaImage-print}{\code{RamigaImage$print()}}
\item \href{#method-RamigaImage-clone}{\code{RamigaImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-info"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-info}{}}}
\subsection{Method \code{info()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-size"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-size}{}}}
\subsection{Method \code{size()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-get_pointer"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-get_pointer}{}}}
\subsection{Method \code{get_pointer()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$get_pointer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
