% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{RamigaMemory}
\alias{RamigaMemory}
\title{TODO}
\value{
TODO

TODO
}
\description{
TODO

TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaEmulator-new}{\code{RamigaMemory$new()}}
\item \href{#method-RamigaEmulator-load_rom}{\code{RamigaMemory$load_rom()}}
\item \href{#method-RamigaEmulator-has_rom}{\code{RamigaMemory$has_rom()}}
\item \href{#method-RamigaEmulator-print}{\code{RamigaMemory$print()}}
\item \href{#method-RamigaEmulator-clone}{\code{RamigaMemory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$new(emulator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emulator}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-load_rom"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-load_rom}{}}}
\subsection{Method \code{load_rom()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$load_rom(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-has_rom"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-has_rom}{}}}
\subsection{Method \code{has_rom()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$has_rom()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaEmulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaEmulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaMemory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
