% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{RamigaOutput}
\alias{RamigaOutput}
\title{The RamigaOutput R6 Class}
\description{
A class to capture output from the Amiga emulator (\link{RamigaEmulator} class).
}
\section{Super class}{
\code{\link[ramiga:RamigaComponent]{ramiga::RamigaComponent}} -> \code{RamigaOutput}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaOutput-capture_frame}{\code{RamigaOutput$capture_frame()}}
\item \href{#method-RamigaOutput-capture_audio_buffer}{\code{RamigaOutput$capture_audio_buffer()}}
\item \href{#method-RamigaOutput-print}{\code{RamigaOutput$print()}}
\item \href{#method-RamigaOutput-clone}{\code{RamigaOutput$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_emulator"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_emulator'><code>ramiga::RamigaComponent$get_emulator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="get_pointer"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-get_pointer'><code>ramiga::RamigaComponent$get_pointer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="initialize"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-initialize'><code>ramiga::RamigaComponent$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ramiga" data-topic="RamigaComponent" data-id="list_options"><a href='../../ramiga/html/RamigaComponent.html#method-RamigaComponent-list_options'><code>ramiga::RamigaComponent$list_options()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-capture_frame"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-capture_frame}{}}}
\subsection{Method \code{capture_frame()}}{
Captures the current pixel information as sent to the monitor of
the virtual device.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$capture_frame(file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A path to a png file where to store the captured frame.}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If file is missing it returns a \code{\link[grDevices:as.raster]{grDevices::as.raster()}} object.
Otherwise, it will save it as png file to the specified path and returns
nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-capture_audio_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-capture_audio_buffer}{}}}
\subsection{Method \code{capture_audio_buffer()}}{
Capture the audio that is currently on the output buffer.
While the emulator is running, audio is continuously generated
and stored in a ring buffer. Use this function to collect data
in this buffer. It can only be read once and the buffer has
a limited capacity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$capture_audio_buffer(file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A file path to store the audio (RIFF wav format).}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If file is omitted, the audio data is returned as
a \code{matrix} of \code{numeric} values. It returns 2 rows (for both
stereo channels). The number of columns correspond with the
number of samples available from the buffer. The waveform
is scaled between -1 and +1, such that it can be played
directly with \code{audio::play()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-print}{}}}
\subsection{Method \code{print()}}{
Prints basic information about the emulator output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
