% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{RamigaOutput}
\alias{RamigaOutput}
\title{The RamigaOutput R6 Class}
\description{
A class to capture output from the Amiga emulator (\link{RamigaEmulator} class).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamigaOutput-new}{\code{RamigaOutput$new()}}
\item \href{#method-RamigaOutput-print}{\code{RamigaOutput$print()}}
\item \href{#method-RamigaOutput-clone}{\code{RamigaOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new Amiga emulator output object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$new(emulator)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-print}{}}}
\subsection{Method \code{print()}}{
Prints basic information about the emulator output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamigaOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamigaOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamigaOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
